# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Astervia_wacraft-core_src_common_model.DescriptiveError:
    properties:
      content:
        description: Additional information about the error provided by the service
          that generated it. E. g. if an ORM returns an error, it will be propagated
          here as the raw error.
      message:
        description: Short text describing the error as best as we can for humans.
        type: string
    type: object
...
  github_com_Astervia_wacraft-core_src_workspace_model.AcceptInvitationRequest:
    properties:
      name:
        description: Required if user doesn't exist
        maxLength: 100
        minLength: 2
        type: string
      password:
        description: Required if user doesn't exist
        maxLength: 72
        minLength: 8
        type: string
      token:
        type: string
    required:
    - token
    type: object
  github_com_Astervia_wacraft-core_src_workspace_model.CreateInvitationRequest:
    properties:
      email:
        type: string
      policies:
        items:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.Policy'
        minItems: 1
        type: array
    required:
    - email
    - policies
    type: object
...
  github_com_Astervia_wacraft-core_src_workspace_model.InvitationResponse:
    properties:
      accepted_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      invited_by:
        type: string
      policies:
        items:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.Policy'
        type: array
      workspace_id:
        type: string
    type: object
  github_com_Astervia_wacraft-core_src_workspace_model.Policy:
    enum:
    - workspace.admin
    - workspace.settings
    - workspace.members
    - phone_config.read
    - phone_config.manage
    - contact.read
    - contact.manage
    - message.read
    - message.send
    - campaign.read
    - campaign.manage
    - campaign.run
    - webhook.read
    - webhook.manage
    type: string
    x-enum-varnames:
    - PolicyWorkspaceAdmin
    - PolicyWorkspaceSettings
    - PolicyWorkspaceMembers
    - PolicyPhoneConfigRead
    - PolicyPhoneConfigManage
    - PolicyContactRead
    - PolicyContactManage
    - PolicyMessageRead
    - PolicyMessageSend
    - PolicyCampaignRead
    - PolicyCampaignManage
    - PolicyCampaignRun
    - PolicyWebhookRead
    - PolicyWebhookManage
...
info:
  contact:
    email: wacraft@astervia.tech
    name: Astervia Dev Team
    url: https://github.com/Astervia
  description: Backend server for the wacraft project. Handles WhatsApp Cloud API
    operations, including message sending, receiving, and webhook handling.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wacraft Server API
  version: 0.1.0
paths:
  /auth/accept-invitation:
    post:
      consumes:
      - application/json
      description: Accepts a workspace invitation. Creates user if doesn't exist.
      parameters:
      - description: Invitation acceptance data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.AcceptInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired invitation
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      summary: Accept invitation
      tags:
      - Auth
...
  /workspace/{workspace_id}/invitation:
    get:
      description: Returns a list of pending invitations for the workspace.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitations
          schema:
            items:
              $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.InvitationResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: List workspace invitations
      tags:
      - Workspace
    post:
      consumes:
      - application/json
      description: Creates an invitation for a user to join the workspace.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Invitation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.CreateInvitationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Invitation created
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_workspace_model.InvitationResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "409":
          description: User already member
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Invite user to workspace
      tags:
      - Workspace
  /workspace/{workspace_id}/invitation/{invitation_id}:
    delete:
      description: Revokes a pending workspace invitation.
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Invitation ID
        in: path
        name: invitation_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Invitation revoked
        "404":
          description: Invitation not found
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Revoke invitation
      tags:
      - Workspace
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"